+++ src/assets/css/Box.css
@@ -1,0 +1,52 @@
+.flex {
+	display: flex;
+}
+
+.required-thing {
+	flex-direction: column;
+	align-items: center;
+	padding: 20px 40px;
+	background: transparent;
+	color: white;
+	border-radius: 10px;
+	transition: all 0.5s;
+	box-shadow: rgba(0, 0, 0, 0.35) 0px 5px 15px;
+	min-width: 144px;
+}
+
+.required-thing:hover {
+	transform: scale(1.5);
+}
+
+.required-things-box {
+	justify-content: space-around;
+	margin: 30px auto;
+	border-radius: 20px;
+	border: 2px black;
+	color: aliceblue;
+}
+
+.required-things-heading {
+	font-size: 25px;
+	margin: 30px 0 47px;
+	font-weight: bold;
+}
+
+@media (max-width: 738px) {
+	.required-things-box {
+		flex-direction: column;
+		align-items: center;
+	}
+
+	.required-thing {
+		margin: 30px auto;
+	}
+
+	.required-thing:last-child {
+		margin-bottom: 100px;
+	}
+
+	.required-thing:hover {
+		transform: scale(1.2);
+	}
+}
